+++ addons/UH60/config/cfgVehicles.hpp
@@ -339,7 +339,7 @@ class CfgVehicles
         class Turrets: Turrets
         {
             #include "turrets\copilot.hpp"
-            #include "turrets\doorguns.hpp"
+            #include "turrets\doorgunsTurnOut.hpp"
         };
         class Exhausts
         {
